[GLM53] KDA state dtype: fp32 recurrent; scatter_states cast on write
Fork's gated_delta_net_state_dtype with cache=auto yields (fp16, fp16)
here, but GLM-5.3 KDA kernels accumulate/write the recurrent state in
fp32: scatter_states then failed with fp16-dest/fp32-src on the first
decode step. Declare (conv fp16, recurrent fp32) and defensively cast
the scatter source to the destination dtype.

+++ vllm/model_executor/layers/mamba/ops/scatter_states.py
@@ -26,4 +26,4 @@ def scatter_states(
     assert src.shape[0] == indices.shape[0]
     assert indices.dtype in (torch.int32, torch.int64)
 
-    state[indices.to(torch.int64)] = src
+    state[indices.to(torch.int64)] = src.to(state.dtype)
